+++ examples/csr-minimal/src/lib.rs
@@ -40,7 +40,7 @@ pub fn LanguageSelector() -> impl IntoView {
                                     name="language"
                                     value=lang
                                     checked=lang.is_active()
-                                    on:click=move |_| i18n.language.set(lang)
+                                    on:click=move |_| lang.activate()
                                 />
                                 <label for=lang>{lang.name}</label>
                             </div>